[1.6][F][B] Add authentication mechanism
At login a JWT Token is created, which is valid for some time
and includes a `group`. This group of the Token is then checked when
accessing a restricted API endpoints. Make it possible to restrict
GraphQL API endpoints to Roles.

At login the JWT token is stored in local storage. Logout is not
implemented yet.

Also this commit includes a private and public key for testing. Should
be remove eventually.

TODO:
- Store JWT token in DB and remove it at logout.
- Implemet Logout Mechanism
- Improve error handling of the GraphQL API. (Hide server error messages)